Moved controllers to directory
